requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
pyyaml==6.0.3
openai==2.30.0
requests==2.33.0
python-dotenv==1.2.2
tavily-python
aiofiles
praw